// src/app/models/product.ts
export interface Product {
    id: number;
    name: string;
    model: string;
    description: string;
    point_desc: string;
    old_price: number;
    price: number;
    image_01: string;
    image_02: string;
    image_03: string;
    category: string;
    sub_category: string;
    availability: string;
    featured: boolean;
}